+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 - Add a general filter control to item lists ([938](../../pull/938))
+- Item list modal dialogs are wider ([939](../../pull/939))
 
 ## 1.16.1